Compile Maxent features to bytecode at runtime

+++ src/main/java/net/gegy1000/earth/server/world/ecology/GrowthIndicator.java
@@ -2,16 +2,16 @@ package net.gegy1000.earth.server.world.ecology;
 
 public interface GrowthIndicator {
     static GrowthIndicator no() {
-        return predictors -> 0.0;
+        return predictors -> 0.0F;
     }
 
-    default GrowthIndicator mul(double scale) {
+    default GrowthIndicator mul(float scale) {
         return predictors -> this.evaluate(predictors) * scale;
     }
 
-    default GrowthIndicator pow(double pow) {
-        return predictors -> Math.pow(this.evaluate(predictors), pow);
+    default GrowthIndicator pow(float pow) {
+        return predictors -> (float) Math.pow(this.evaluate(predictors), pow);
     }
 
-    double evaluate(GrowthPredictors predictors);
+    float evaluate(GrowthPredictors predictors);
 }
